+++ src/features/posts/model/index.ts
@@ -3,3 +3,4 @@ export { useMutationDeletePost } from "./useMutationDeletePost";
 export { useMutationUpdatePost } from "./useMutationUpdatePost";
 export { useSuspenseQueryGetPosts } from "./useSuspenseQueryGetPosts";
 export { useSuspenseQueryGetTags } from "./useSuspenseQueryGetTags";
+export { usePost } from "./usePost";
