+++ src/features/users/index.ts
@@ -1,1 +1,2 @@
+export { UserProfile } from "./ui";
 export * from "./model";
